Mod: Moved ContentBuilder to the Common namespace.
