+++ NPLEditor/Program.cs
@@ -1,4 +1,5 @@
 ﻿using NPLEditor;
+using NPLEditor.Common;
 using NPLEditor.Data;
 using Serilog;
 using System;
